requirements.txt
fastapi[standard]
uvicorn
joblib
scikit-learn
mlflow
kafka-python
eurybia